requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
vaderSentiment==3.3.2
pydantic==2.8.2
pymongo
motor
python-dotenv
fer
opencv-python
numpy
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
openai>=1.40.0
email-validator>=2.0.0
python-multipart>=0.0.9
